requirements.txt
google-auth==2.0.2
google-auth-oauthlib==0.4.6
google-auth-httplib2==0.1.0
google-api-python-client==2.27.0
tzlocal==4.2
selenium==4.10.0
webdriver-manager